Python-CODING
coding tests and stuff on python, i am only in 8th grade so i apologize for bad code, i'm trying to right a calculator just to see what i can do. Credit to github user "Bruh-Not" for not only being my classmate and helping with some of this, but making a base for some of these projects. such as "Calculator"